+++ lib/linkify-it-rb/re.rb
@@ -32,11 +32,13 @@ module LinkifyRe
 
     SRC_PORT  = '(?::(?:6(?:[0-4]\\d{3}|5(?:[0-4]\\d{2}|5(?:[0-2]\\d|3[0-5])))|[1-5]?\\d{1,4}))?'
 
-    SRC_HOST_TERMINATOR = '(?=$|' + TEXT_SEPARATORS + '|' + SRC_Z_P_CC + ')(?!-|_|:\\d|\\.-|\\.(?!$|' + SRC_Z_P_CC + '))'
+    # moved SRC_HOST_TERMINATOR into re_src_host_terminator
 
     # moved SRC_PATH into re_src_path
 
-    SRC_EMAIL_NAME  = '[\\-;:&=\\+\\$,\\"\\.a-zA-Z0-9_]+'
+    # Allow anything in markdown spec, forbid quote (") at the first position
+    # because emails enclosed in quotes are far more common
+    SRC_EMAIL_NAME  = '[\\-;:&=\\+\\$,\\.a-zA-Z0-9_][\\-;:&=\\+\\$,\\"\\.a-zA-Z0-9_]*'
     SRC_XN          = 'xn--[a-z0-9\\-]{1,59}'
 
     # More to read about domain names
@@ -56,10 +58,7 @@ module LinkifyRe
         '|' +
         '(?:' + SRC_PSEUDO_LETTER + ')' +
         '|' +
-        # don't allow `--` in domain names, because:
-        # - that can conflict with markdown &mdash; / &ndash;
-        # - nobody use those anyway
-        '(?:' + SRC_PSEUDO_LETTER + '(?:-(?!-)|' + SRC_PSEUDO_LETTER + '){0,61}' + SRC_PSEUDO_LETTER + ')' +
+        '(?:' + SRC_PSEUDO_LETTER + '(?:-|' + SRC_PSEUDO_LETTER + '){0,61}' + SRC_PSEUDO_LETTER + ')' +
       ')'
 
     SRC_HOST =
@@ -80,23 +79,16 @@ module LinkifyRe
     TPL_HOST_NO_IP_FUZZY =
       '(?:(?:(?:' + SRC_DOMAIN + ')\\.)+(?:%TLDS%))'
 
-    SRC_HOST_STRICT                   = SRC_HOST + SRC_HOST_TERMINATOR
-    TPL_HOST_FUZZY_STRICT             = TPL_HOST_FUZZY + SRC_HOST_TERMINATOR
-    SRC_HOST_PORT_STRICT              = SRC_HOST + SRC_PORT + SRC_HOST_TERMINATOR
-    TPL_HOST_PORT_FUZZY_STRICT        = TPL_HOST_FUZZY + SRC_PORT + SRC_HOST_TERMINATOR
-    TPL_HOST_PORT_NO_IP_FUZZY_STRICT  = TPL_HOST_NO_IP_FUZZY + SRC_PORT + SRC_HOST_TERMINATOR
-
     #------------------------------------------------------------------------------
     # Main rules
 
     # Rude test fuzzy links by host, for quick deny
     TPL_HOST_FUZZY_TEST = 'localhost|www\\.|\\.\\d{1,3}\\.|(?:\\.(?:%TLDS%)(?:' + SRC_Z_P_CC + '|>|$))'
-    TPL_EMAIL_FUZZY     = '(^|' + TEXT_SEPARATORS + '|\\(|' +SRC_Z_CC + ')(' + SRC_EMAIL_NAME + '@' + TPL_HOST_FUZZY_STRICT + ')'
 
     # moved TPL_LINK_FUZZY and TPL_LINK_NO_IP_FUZZY into build_re
 
   #------------------------------------------------------------------------------
-  def build_re(opts)
+  def build_re(opts = {})
     re = {
       src_Any:              SRC_ANY,
       src_Cc:               SRC_CC,
@@ -108,7 +100,7 @@ module LinkifyRe
       src_ip4:              SRC_IP4,
       src_auth:             SRC_AUTH,
       src_port:             SRC_PORT,
-      src_host_terminator:  SRC_HOST_TERMINATOR,
+      src_host_terminator:  re_src_host_terminator(opts),
       src_path:             re_src_path(opts),
       src_email_name:       SRC_EMAIL_NAME,
       src_xn:               SRC_XN,
@@ -118,60 +110,72 @@ module LinkifyRe
 
       tpl_host_fuzzy:                   TPL_HOST_FUZZY,
       tpl_host_no_ip_fuzzy:             TPL_HOST_NO_IP_FUZZY,
-      src_host_strict:                  SRC_HOST_STRICT,
-      tpl_host_fuzzy_strict:            TPL_HOST_FUZZY_STRICT,
-      src_host_port_strict:             SRC_HOST_PORT_STRICT,
-      tpl_host_port_fuzzy_strict:       TPL_HOST_PORT_FUZZY_STRICT,
-      tpl_host_port_no_ip_fuzzy_strict: TPL_HOST_PORT_NO_IP_FUZZY_STRICT,
+      src_host_strict:                  SRC_HOST + re_src_host_terminator(opts),
+      tpl_host_fuzzy_strict:            TPL_HOST_FUZZY + re_src_host_terminator(opts),
+      src_host_port_strict:             SRC_HOST + SRC_PORT + re_src_host_terminator(opts),
+      tpl_host_port_fuzzy_strict:       TPL_HOST_FUZZY + SRC_PORT + re_src_host_terminator(opts),
+      tpl_host_port_no_ip_fuzzy_strict: TPL_HOST_NO_IP_FUZZY + SRC_PORT + re_src_host_terminator(opts),
 
-      tpl_host_fuzzy_test:  TPL_HOST_FUZZY_TEST,
-      tpl_email_fuzzy:      TPL_EMAIL_FUZZY
+      tpl_host_fuzzy_test:  TPL_HOST_FUZZY_TEST
     }
+
+    re[:tpl_email_fuzzy] =
+      '(^|' + TEXT_SEPARATORS + '|"|\\(|' + SRC_Z_CC + ')' +
+      '(' + SRC_EMAIL_NAME + '@' + re[:tpl_host_fuzzy_strict] + ')'
 
     # Fuzzy link can't be prepended with .:/\- and non punctuation.
     # but can start with > (markdown blockquote)
     re[:tpl_link_fuzzy] =
         '(^|(?![.:/\\-_@])(?:[$+<=>^`|\uff5c]|' + SRC_Z_P_CC + '))' +
-        '((?![$+<=>^`|\uff5c])' + TPL_HOST_PORT_FUZZY_STRICT + re[:src_path] + ')'
+        '((?![$+<=>^`|\uff5c])' + re[:tpl_host_port_fuzzy_strict] + re[:src_path] + ')'
 
     # Fuzzy link can't be prepended with .:/\- and non punctuation.
     # but can start with > (markdown blockquote)
     re[:tpl_link_no_ip_fuzzy] =
       '(^|(?![.:/\\-_@])(?:[$+<=>^`|\uff5c]|' + SRC_Z_P_CC + '))' +
-      '((?![$+<=>^`|\uff5c])' + TPL_HOST_PORT_NO_IP_FUZZY_STRICT + re[:src_path] + ')'
+      '((?![$+<=>^`|\uff5c])' + re[:tpl_host_port_no_ip_fuzzy_strict] + re[:src_path] + ')'
 
     return re
   end
 
   #------------------------------------------------------------------------------
-  def re_src_path(opts = nil)
+  def re_src_path(opts = {})
     '(?:' +
       '[/?#]' +
         '(?:' +
-          '(?!' + SRC_Z_CC + '|' + TEXT_SEPARATORS + '|[()\\[\\]{}.,"\'?!\\-]).|' +
+          '(?!' + SRC_Z_CC + '|' + TEXT_SEPARATORS + '|[()\\[\\]{}.,"\'?!\\-;]).|' +
           '\\[(?:(?!' + SRC_Z_CC + '|\\]).)*\\]|' +
           '\\((?:(?!' + SRC_Z_CC + '|[)]).)*\\)|' +
           '\\{(?:(?!' + SRC_Z_CC + '|[}]).)*\\}|' +
           '\\"(?:(?!' + SRC_Z_CC + '|["]).)+\\"|' +
           "\\'(?:(?!" + SRC_Z_CC + "|[']).)+\\'|" +
-          "\\'(?=" + SRC_PSEUDO_LETTER + '|[-]).|' +  # allow `I'm_king` if no pair found
-          '\\.{2,3}[a-zA-Z0-9%/]|' + # github has ... in commit range links. Restrict to
+          "\\'(?=" + SRC_PSEUDO_LETTER + '|[-])|' +  # allow `I'm_king` if no pair found
+          '\\.{2,}[a-zA-Z0-9%/&]|' + # google has many dots in "google search" links (#66, #81).
+                                     # github has ... in commit range links,
+                                     # Restrict to
                                      # - english
                                      # - percent-encoded
                                      # - parts of file path
+                                     # - params separator
                                      # until more examples found.
-          '\\.(?!' + SRC_Z_CC + '|[.]).|' +
+          '\\.(?!' + SRC_Z_CC + '|[.]|$)|' +
           (opts && opts[:'---'] ?
             '\\-(?!--(?:[^-]|$))(?:-*)|'  # `---` => long dash, terminate
           :
             '\\-+|'
           ) +
-          '\\,(?!' + SRC_Z_CC + ').|' +      # allow `,,,` in paths
-          '\\!(?!' + SRC_Z_CC + '|[!]).|' +
-          '\\?(?!' + SRC_Z_CC + '|[?]).' +
+          '\\,(?!' + SRC_Z_CC + '|$)|' +      # allow `,,,` in paths
+          '\\;(?!' + SRC_Z_CC + '|$)|' +      # allow `;` if not followed by space-like char
+          '\\!+(?!' + SRC_Z_CC + '|[!]|$)|' + # allow `!!!` in paths, but not at the end
+          '\\?(?!' + SRC_Z_CC + '|[?]|$)' +
         ')+' +
       '|\\/' +
     ')?'
-end
+  end
 
+  #------------------------------------------------------------------------------
+  def re_src_host_terminator(opts = {})
+    '(?=$|' + TEXT_SEPARATORS + '|' + SRC_Z_P_CC + ')' +
+    '(?!' + (opts['---'] ? '-(?!--)|' : '-|') + '_|:\\d|\\.-|\\.(?!$|' + SRC_Z_P_CC + '))'
+  end
 end