+++ rubymotion/spec/linkify-it-rb/test_spec.rb
@@ -273,5 +273,39 @@ describe 'API' do
     l = Linkify.new(nil, { :'---' => true })
 
     expect(l.match('http://e.com/foo---bar')[0].text).to eq 'http://e.com/foo'
+    expect(l.match('text@example.com---foo')).to be_empty
+  end
+
+  #------------------------------------------------------------------------------
+  it 'should find a match at the start' do
+    l = Linkify.new
+
+    l.set({ fuzzyLink: true })
+
+    expect(l.matchAtStart('http://google.com 123').text).to eq 'http://google.com'
+    expect(l.matchAtStart('google.com 123')).to be_nil
+    expect(l.matchAtStart('  http://google.com 123')).to be_nil
+  end
+
+  #------------------------------------------------------------------------------
+  it 'matchAtStart should not interfere with normal match' do
+    l = Linkify.new
+
+    str = 'http://google.com http://google.com'
+    expect(l.matchAtStart(str)).not_to be_nil
+    expect(l.match(str).length).to eq 2
+
+    str = 'aaa http://google.com http://google.com'
+    expect(l.matchAtStart(str)).to be_nil
+    expect(l.match(str).length).to eq 2
+  end
+
+  #------------------------------------------------------------------------------
+  it 'should not match incomplete links' do
+    # regression test for https://github.com/markdown-it/markdown-it/issues/868
+    l = Linkify.new
+
+    expect(l.matchAtStart('http://')).to be_nil
+    expect(l.matchAtStart('https://')).to be_nil
   end
 end