+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+4.0.1
+-------
+
+Synced with linkify-it 4.0.1, see the [CHANGELOG](https://github.com/markdown-it/linkify-it/blob/master/CHANGELOG.md)
+
+3.0.3
+-------
+
+Synced with linkify-it 3.0.3, see the [CHANGELOG](https://github.com/markdown-it/linkify-it/blob/master/CHANGELOG.md)
+
 2.0.3
 -------